+++ ray_julia_jll/src/common.jl
@@ -1,24 +1,3 @@
-using CxxWrap
-using CxxWrap.StdLib: StdVector, SharedPtr
-using Serialization
-using libcxxwrap_julia_jll
-
-JLLWrappers.@generate_wrapper_header("ray_julia")
-JLLWrappers.@declare_library_product(ray_julia, "julia_core_worker_lib.so")
-@wrapmodule(joinpath(artifact"ray_julia", "julia_core_worker_lib.so"))
-
-function __init__()
-    JLLWrappers.@generate_init_header(libcxxwrap_julia_jll)
-    JLLWrappers.@init_library_product(
-        ray_julia,
-        "julia_core_worker_lib.so",
-        RTLD_GLOBAL,
-    )
-
-    JLLWrappers.@generate_init_footer()
-    @initcxx
-end  # __init__()
-
 const STATUS_CODE_SYMBOLS = (:OK,
                              :OutOfMemory,
                              :KeyError,
